requirements.txt
mysql-connector-python
requests~=2.24.0
Flask~=1.1.2
peewee